Create internal `gempyor._click` module
Extracted `DurationParamType` and `MemoryParamType` out of
`gempyor.shared_cli` into `gempyor._click`.
