+++ flepimop/gempyor_pkg/tests/_click/test_duration_param_type_class.py
@@ -1,10 +1,10 @@
 from datetime import timedelta
-from typing import Any, Literal
+from typing import Literal
 
 from click.exceptions import BadParameter
 import pytest
 
-from gempyor.shared_cli import DurationParamType
+from gempyor._click import DurationParamType
 
 
 @pytest.mark.parametrize("nonnegative", (True, False))
